Refactoring NoPrefixForInterfacesRuleTest
