+++ src/test/java/com/societegenerale/aut/test/TotallyGoodInterfaceName.java
@@ -1,4 +1,4 @@
-package com.societegenerale.aut.main;
+package com.societegenerale.aut.test;
 
 public interface TotallyGoodInterfaceName {
 }
